Move sort_to_local_joint_order to trajectory_operations
