+++ joint_trajectory_controller/include/joint_trajectory_controller/trajectory.hpp
@@ -163,38 +163,6 @@ private:
 };
 
 /**
- * \return The map between \p t1 indices (implicitly encoded in return vector indices) to \p t2
- * indices. If \p t1 is <tt>"{C, B}"</tt> and \p t2 is <tt>"{A, B, C, D}"</tt>, the associated
- * mapping vector is <tt>"{2, 1}"</tt>.
- */
-template <class T>
-inline std::vector<size_t> mapping(const T & t1, const T & t2)
-{
-  // t1 must be a subset of t2
-  if (t1.size() > t2.size())
-  {
-    return std::vector<size_t>();
-  }
-
-  std::vector<size_t> mapping_vector(t1.size());  // Return value
-  for (auto t1_it = t1.begin(); t1_it != t1.end(); ++t1_it)
-  {
-    auto t2_it = std::find(t2.begin(), t2.end(), *t1_it);
-    if (t2.end() == t2_it)
-    {
-      return std::vector<size_t>();
-    }
-    else
-    {
-      const size_t t1_dist = std::distance(t1.begin(), t1_it);
-      const size_t t2_dist = std::distance(t2.begin(), t2_it);
-      mapping_vector[t1_dist] = t2_dist;
-    }
-  }
-  return mapping_vector;
-}
-
-/**
  * \param current_position The current position given from the controller, which will be adapted.
  * \param next_position Next position from which to compute the wraparound offset, i.e.,
  *      the first trajectory point
